Move `createExecutor` to `Http2TestServer`
