+++ test/jdk/java/net/httpclient/lib/jdk/httpclient/test/lib/http2/Http2TestServer.java
@@ -28,6 +28,9 @@ import java.net.*;
 import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.concurrent.ThreadFactory;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.Predicate;
@@ -193,7 +196,16 @@ public class Http2TestServer implements AutoCloseable {
 
     private static ExecutorService createExecutor(String name) {
         String threadNamePrefix = "http2-test-server-worker" + (name != null ? "[%s]".formatted(name) : "");
-        return HttpServerAdapters.createExecutor(threadNamePrefix);
+        ThreadFactory threadFactory = createThreadFactory(threadNamePrefix);
+        return Executors.newCachedThreadPool(threadFactory);
+    }
+
+    private static ThreadFactory createThreadFactory(String threadNamePrefix) {
+        AtomicInteger threadCounter = new AtomicInteger();
+        return (Runnable task) -> {
+            String name = "%s-%d".formatted(threadNamePrefix, threadCounter.incrementAndGet());
+            return new Thread(task, name);
+        };
     }
 
     /**
